QT-11:Fix sql syntax in spring boot jpa

+++ src/main/java/com/spring/app/customers/repository/ForControlRepository.java
@@ -3,5 +3,5 @@ package com.spring.app.customers.repository;
 import com.spring.app.customers.models.ForControl;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ForControlRepository extends JpaRepository<ForControl, Integer> {
+public interface ForControlRepository extends JpaRepository<ForControl, Long> {
 }
